Adds a test for date period setting.
